Finished Doc for User
Made a short class description as well as documentation to all externly used methods.

May be the documentation needs a clean up. I havent figured out how to make clean
reference to a methode.

But all together it fullfils the purpose of informing.

Note: I figured out, that you can look up the doc of anything by clicking on the name
and pressing crlt-q.

	modified:   app/models/user.rb